First Upload of Clean src/ Files
I decided it would be a good start for the repo if I uploaded the
untouched src directory and we can make changes from there. If you
guys think we need the other project files we can re-work things or
I can just upload them. Hope this works for everyone, thank you.