Added code to find the selected method [Read Description]
I added a doubleclick listener to find what method is clicked on
by a plug in user. As it stands right now, when the plug in is loaded
you must click on the MethodIndices Tab and then click the 'Start Indexing'
button and wait for the total number of methods to be displayed. Then you
can double click on a method in the package explorer and the indexed
value will be pulled from the hashmap created by the button press and
displayed in the MethodIndicesView. There is still a tokenize bug you
will see but it is working finally. I tried my best to catch all exceptions
caused by erroneous double clicks.